# src/swagger.yaml
openapi: 3.0.0
info:
  title: WhatsApp Bridge API
  description: A RESTful API service that bridges WhatsApp Web with custom applications
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
...
    Subscription:
      type: object
      required:
        - url
        - events
        - secret
      properties:
        url:
          type: string
          description: Webhook URL to receive events
        events:
          type: array
          items:
            $ref: '#/components/schemas/AvailableEvents'
          description: List of events to subscribe to
        secret:
          type: string
          description: Secret key used to authenticate webhook calls
...
    AvailableEvents:
      type: string
      enum:
        - message
        - message_create
        - message_ack
        - group_join
        - group_leave
        - group_update
      description: Available events for subscription
...
security:
  - BearerAuth: []
paths:
...
  /subscriptions:
    post:
      tags:
        - Webhooks
      summary: Create webhook subscription
      description: Subscribe to WhatsApp events with a webhook URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Subscription'
      responses:
        201:
          description: Subscription created successfully
        400:
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
        - Webhooks
      summary: List all subscriptions
      description: Get a list of all active webhook subscriptions
      responses:
        200:
          description: List of active subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    url:
                      type: string
                    events:
                      type: array
                      items:
                        $ref: '#/components/schemas/AvailableEvents'
    delete:
      tags:
        - Webhooks
      summary: Delete a subscription
      description: Remove a webhook subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
              properties:
                url:
                  type: string
                  description: URL of the subscription to delete
      responses:
        200:
          description: Subscription removed successfully
        404:
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...